Adds translation support and addt styles
