+++ src/components/formApp.scss
@@ -27,3 +27,9 @@
   font-size: 28px;
   line-height: 40px;
 }
+
+#question-text2 {
+  font-size: 16px;
+  line-height: 26px;
+  font-weight: 100;
+}
