+++ src/components/results.scss
